+++ CHANGELOG.md
@@ -33,6 +33,7 @@
 * Add error reporting with Sentry support on client (default ErrorBoundary), server and Redux middleware @sneridagh
 * Review chunks policy. Keep it in sync with Webpack 4 policy with entrypoint bundles @sneridagh
 * Added tiles reducers @robgietema
+* Merged block styling to inline toolbar @robgietema
 
 ## 0.4.0 (2017-05-03)